requirements.txt
python-dateutil==2.5.3
requests==2.10.0
Flask==0.10.1
elasticsearch==5.4.0
